# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Software Engineering Course Project",
        "title": "SE Project",
        "contact": {
            "name": "Mahdieh Moghiseh",
            "email": "mahdiehmoghiseh81@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/trainee/profile/{id}": {
            "get": {
                "description": "Retrieves the profile information of a trainee by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trainee"
                ],
                "summary": "Get trainee profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the user",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trainee profile information",
                        "schema": {
                            "$ref": "#/definitions/dto.TraineeResponse"
                        }
                    },
                    "404": {
                        "description": "Trainee not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates the profile information of a trainee by UserID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trainee"
                ],
                "summary": "Edit trainee profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the user",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Trainee profile data",
                        "name": "trainee",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TraineeEdit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated trainee profile",
                        "schema": {
                            "$ref": "#/definitions/dto.TraineeResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Trainee not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/trainee/request/": {
            "post": {
                "description": "Creates a new program request with the provided data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trainee"
                ],
                "summary": "Create program request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the user",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Program request data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ProgramRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created program request",
                        "schema": {
                            "$ref": "#/definitions/dto.ProgramRequest"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Invalid user ID or not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/trainee/request/{id}": {
            "get": {
                "description": "Retrieves the request of a trainee by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trainee"
                ],
                "summary": "Get trainee request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the user",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trainee request",
                        "schema": {
                            "$ref": "#/definitions/dto.RequestsInTrainerPage"
                        }
                    },
                    "404": {
                        "description": "Trainee not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Change request status by trainee",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trainee"
                ],
                "summary": "Change request status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the user",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Request Change Status",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TraineeChangeStatus"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trainee Change Status",
                        "schema": {
                            "$ref": "#/definitions/dto.ProgramRequestSetPrice"
                        }
                    },
                    "404": {
                        "description": "Trainee not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ProgramRequest": {
            "type": "object",
            "required": [
                "active_days",
                "trainerID"
            ],
            "properties": {
                "active_days": {
                    "type": "array",
                    "items": {
                        "type": "boolean"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "traineeID": {
                    "type": "integer"
                },
                "trainerID": {
                    "type": "integer"
                }
            }
        },
        "dto.ProgramRequestSetPrice": {
            "type": "object",
            "required": [
                "traineeID",
                "trainerID"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "traineeID": {
                    "type": "integer"
                },
                "trainerID": {
                    "type": "integer"
                }
            }
        },
        "dto.RequestsInTrainerPage": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "trainee_name": {
                    "type": "string"
                }
            }
        },
...
        "dto.TraineeChangeStatus": {
            "type": "object",
            "properties": {
                "request_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dto.TraineeEdit": {
            "type": "object",
            "required": [
                "active_days",
                "contact",
                "country",
                "height",
                "language",
                "medicalHistory",
                "sports",
                "status",
                "user_name",
                "weight"
            ],
            "properties": {
                "active_days": {
                    "type": "array",
                    "items": {
                        "type": "boolean"
                    }
                },
                "contact": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "language": {
                    "type": "string"
                },
                "medicalHistory": {
                    "type": "string"
                },
                "sports": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/dto.UserEdit"
                },
                "user_name": {
                    "type": "string"
                },
                "weight": {
                    "type": "integer"
                }
            }
        },
...
        "dto.TraineeProfileCard": {
            "type": "object",
            "properties": {
                "contact": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                },
                "wallet": {
                    "type": "integer"
                }
            }
        },
        "dto.TraineeResponse": {
            "type": "object",
            "properties": {
                "health_problems": {
                    "type": "string"
                },
                "sport_experience": {
                    "type": "string"
                },
                "trainee_profile_card": {
                    "$ref": "#/definitions/dto.TraineeProfileCard"
                }
            }
        },
...
        "dto.UserEdit": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "block": {
                    "type": "boolean"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "info_id": {
                    "type": "string"
                },
                "info_type": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "wallet": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}